[QuantityValueRange] Bugfix for min or max value with decimal places.
If you entered a number with decimal places in a QuantitValueRange field
and try to save the object, you'll get a warning

  `Argument #3 ($step) must not exceed the specified range in php shell code`

since the default of 1 is too large. Additionally, a step fit for the numbers
entered could not be passed to `getValue()`/`getRange()` due the limitation to `int`.

With this fix the step necessary for the numbers in use is determined automatically,
and you could enter your own step as specified with php's range function, see
https://www.php.net/manual/en/function.range.php